Add launch file and alphabetize
